Bump dependencies where possible
The only dependency that is not upgraded is `git2go`, because the
latest version is incompatible with the current version in our repo.

Built a few packages to verify nothing is breaking. I've checked the
changelogs for each upgrade and verified that they either have no
breaking changes or have a few that doesn't affect `solbuild`.

Signed-off-by: Gavin Zhao <git@gzgz.dev>